Attempt to fix covr error
Use the github version of covr (3.5.1.9003) to try to fix
https://github.com/leekgroup/recount/runs/3286906239?check_suite_focus=true#step:26:27 (covr 3.5.1)

https://github.com/r-lib/covr/commit/ddbfbec356b9d985368131e55bc860463ba09a47
implements a fix to https://github.com/r-lib/covr/issues/420 which maybe
explains this issue

Using the Bioconductor devel docker I cannot reproduce the error on the GHA
run using the same docker. That's when executing the test code manually, not
with covr (I'm still in the process of running R CMD check + covr there).